refactor(state): move the firmware cells into their own module
`Local` and `LocalCell` are the building blocks every module will use once
state moves out of `state.rs`, so they get a module of their own that
documents the single-hart invariant once.
